4е задание
